+++ TransportManagementSystem/Engine.cs
@@ -5,13 +5,13 @@ using System.Text;
 using System.Threading.Tasks;
 using TransportManagementSystem.Contracts;
 
-namespace TransportManagementSystem.Transport
+namespace TransportManagementSystem
 {
     public class Engine : IRepairable
     {
         public string Model { get; set; }
         public string Brand { get; set; }
         public int HorsePower { get; set; }
-        public DateTime? LastRepairDate { get; set; }
+
     }
 }
